Add support for responding to interactions using BaseInteractions
